Updated the components folder
