--- src/components/BookForm.js
@@ -1,43 +1,0 @@
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
-
-const BookForm = ({ onAdd }) => {
-  const [title, setTitle] = useState('');
-  const [author, setAuthor] = useState('');
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (title && author) {
-      const newBook = { title, author, id: new Date().getTime() };
-      onAdd(newBook);
-      setTitle('');
-      setAuthor('');
-    }
-  };
-
-  return (
-    <form onSubmit={handleSubmit} className="book-form">
-      <input
-        type="text"
-        placeholder="Title"
-        value={title}
-        onChange={(e) => setTitle(e.target.value)}
-        className="book-form-input"
-      />
-      <input
-        type="text"
-        placeholder="Author"
-        value={author}
-        onChange={(e) => setAuthor(e.target.value)}
-        className="book-form-input"
-      />
-      <button type="submit" className="book-form-button">Add Book</button>
-    </form>
-  );
-};
-
-BookForm.propTypes = {
-  onAdd: PropTypes.func.isRequired,
-};
-
-export default BookForm;
