+++ src/index.js
@@ -10,5 +10,4 @@ root.render(
   <Provider store={store}>
     <App />
   </Provider>,
-  document.getElementById('root'),
 );
