+++ src/redux/categories/categoriesSlice.js
@@ -1,13 +1,20 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-const initialState = {
-  categories: ['Under construction'],
-};
-
 const categoriesSlice = createSlice({
   name: 'categories',
-  initialState,
-  reducers: {},
+  initialState: {
+    categories: [],
+  },
+  reducers: {
+    checkStatus: (state, action) => {
+      // eslint-disable-next-line no-param-reassign
+      state.categories = action.payload === 'Under construction'
+        ? 'Under construction'
+        : state.categories;
+    },
+  },
 });
+
+export const { checkStatus } = categoriesSlice.actions;
 
 export default categoriesSlice.reducer;
